remove global ags
