+++ src/service/battery.ts
@@ -1,5 +1,5 @@
 import Gio from 'gi://Gio';
-import Service from './service.js';
+import Service from '../service.js';
 import { timeout } from '../utils.js';
 import { loadInterfaceXML } from '../utils.js';
 import { type BatteryProxy } from '../dbus/types.js';
@@ -13,7 +13,7 @@ const DeviceState = {
     FULLY_CHARGED: 4,
 };
 
-class BatteryService extends Service {
+class Battery extends Service {
     static {
         Service.register(this, {
             'closed': [],
@@ -80,20 +80,4 @@ class BatteryService extends Service {
     }
 }
 
-export default class Battery {
-    static _instance: BatteryService;
-
-    static get instance() {
-        Service.ensureInstance(Battery, BatteryService);
-        return Battery._instance;
-    }
-
-    static get available() { return Battery.instance.available; }
-    static get percent() { return Battery.instance.percent; }
-    static get charging() { return Battery.instance.charging; }
-    static get charged() { return Battery.instance.charged; }
-
-    static get iconName() { return Battery.instance.icon_name; }
-    static get icon_name() { return Battery.instance.icon_name; }
-    static get ['icon-name']() { return Battery.instance.icon_name; }
-}
+export default new Battery();
