+++ src/widget.ts
@@ -39,23 +39,3 @@ export const Revealer = (args: object) => constructor(AgsRevealer, args);
 export const Scrollable = (args: object) => constructor(AgsScrollable, args);
 export const Slider = (args: object) => constructor(AgsSlider, args);
 export const Stack = (args: object) => constructor(AgsStack, args);
-
-// so it is still in global scope through ags.Widget
-Widget.Widget = Widget;
-Widget.Box = Box;
-Widget.Button = Button;
-Widget.CenterBox = CenterBox;
-Widget.CircularProgress = CircularProgress;
-Widget.Entry = Entry;
-Widget.EventBox = EventBox;
-Widget.Icon = Icon;
-Widget.Label = Label;
-Widget.Menu = Menu;
-Widget.MenuItem = MenuItem;
-Widget.Overlay = Overlay;
-Widget.ProgressBar = ProgressBar;
-Widget.Revealer = Revealer;
-Widget.Scrollable = Scrollable;
-Widget.Slider = Slider;
-Widget.Stack = Stack;
-Widget.Window = Window;
